backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.2",
    "google-generativeai>=0.8.6",
    "pypdf>=6.9.2",
    "python-multipart>=0.0.22",
    "uvicorn>=0.42.0",
]

[dependency-groups]
dev = [
    "fpdf2>=2.8.7",
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.1.0",
]
